created app.py
This file is similar to the one in PhaseOne, except that I have to partition the keys and distribute them evenly across N number of nodes. Any Node can recieve a request from a client for a key. If that key is on that node, then that can service the request. If the key is on a different node, then the rquest needs to be forwarded to the correct node